python-zipper/zipper/requirements.txt
django
boto3
django-storages
zipstream
redis
gunicorn
django-health-check
